% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-gantt.R
\name{geom_gantt2}
\alias{geom_gantt2}
\title{A gantt plot}
\usage{
geom_gantt2(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{add_line:}{If FALSE, the default, add a dashed line for each rectangle.}

\item{linecolor:}{Color of the dashed line, The default color is black.}

\item{point_size:}{Size of middle point, default is 2pt.}

\item{flip:}{Whether to flip x and y, defualt is FALSE, if you add the coord_flip, please set flip = TRUE.}
}
\description{
The gantt plot uses bar charts to show the intrinsic relationship of
project, schedule, and other time-dependent system progress over time.
This function requires a processed wider data as input. See Examples
for details.
}
\examples{
# You can use head(gantt_data_wider) to view the data type of the
input data.
# head(gantt_data_wider)
ggplot(gantt_data_wider)+
  geom_gantt2(aes(x = CancerType,
                  minimum = low, maximum = high,
                  middle = middle, fill = group),
              position = position_dodge(0.2),
              # stroke and color of meddle points
              stroke = 0.5, point_size = 2, color = "black",
              # widths of the rectangles
              width = 0.1,
              # add dashed line
              add_line = T
  )
}
