% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGantt-package.R
\docType{package}
\name{ggGantt-package}
\alias{ggGantt}
\alias{ggGantt-package}
\title{ggGantt: Make a gantt plot}
\description{
The gantt plot uses bar charts to show the intrinsic relationship of project, schedule, and other time-dependent system progress over time.
}
\keyword{internal}
